Ошибка по библиотеке в Bin/Debag Не понимаю что не так... уже все перепробовал, даже АВ отключал...
