--- CS_Basics_Lesson7/Lesson7_Doubler/Lesson7_Doubler/Lesson7_Doubler.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson7_Doubler", "Lesson7_Doubler\Lesson7_Doubler.csproj", "{498632A9-8CAA-423F-A46D-0A1EDA02B59B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{498632A9-8CAA-423F-A46D-0A1EDA02B59B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{498632A9-8CAA-423F-A46D-0A1EDA02B59B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{498632A9-8CAA-423F-A46D-0A1EDA02B59B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{498632A9-8CAA-423F-A46D-0A1EDA02B59B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {070124E7-D1BB-4474-BACA-ED139812E0D5}
-	EndGlobalSection
-EndGlobal
