+++ CS_Basics_Lesson7/Lesson7_WF_Udvoitel/WF_Udvoitel/WF_Udvoitel.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31205.134
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WF_Udvoitel", "WF_Udvoitel\WF_Udvoitel.csproj", "{9E114994-BF20-4C1C-8A87-D9FAFFFD38F1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9E114994-BF20-4C1C-8A87-D9FAFFFD38F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9E114994-BF20-4C1C-8A87-D9FAFFFD38F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9E114994-BF20-4C1C-8A87-D9FAFFFD38F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9E114994-BF20-4C1C-8A87-D9FAFFFD38F1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {10B3923C-6D33-46B4-A1E0-2829FC946803}
+	EndGlobalSection
+EndGlobal
