+++ CS_Basics_Lesson7/Lesson7_WF_Udvoitel/WF_Udvoitel/WF_Udvoitel/bin/Debug/netcoreapp3.1/WF_Udvoitel.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v3.1",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v3.1": {
+      "WF_Udvoitel/1.0.0": {
+        "runtime": {
+          "WF_Udvoitel.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "WF_Udvoitel/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
