+++ CS_Basics_Lesson7/Lesson7_WF_Udvoitel/WF_Udvoitel/WF_Udvoitel/bin/Debug/netcoreapp3.1/WF_Udvoitel.runtimeconfig.dev.json
@@ -1,0 +1,9 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\kozyr\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\kozyr\\.nuget\\packages",
+      "C:\\Program Files (x86)\\Microsoft Visual Studio\\Shared\\NuGetPackages"
+    ]
+  }
+}
