+++ CS_Basics_Lesson7/Lesson7_WF_Udvoitel/WF_Udvoitel/WF_Udvoitel/bin/Debug/netcoreapp3.1/WF_Udvoitel.runtimeconfig.json
@@ -1,0 +1,9 @@
+{
+  "runtimeOptions": {
+    "tfm": "netcoreapp3.1",
+    "framework": {
+      "name": "Microsoft.WindowsDesktop.App",
+      "version": "3.1.0"
+    }
+  }
+}
